refactor: buttontext 컴포넌트 삭제 및 buttonmain으로 통일 #12
