+++ src/components/buttons/index.ts
@@ -1,6 +1,5 @@
 import ButtonMain from './ButtonMain';
-import ButtonText from './ButtonText';
 import ButtonCheckbox from './ButtonCheckbox';
 import ButtonAssist from './ButtonAssist';
 
-export { ButtonMain, ButtonText, ButtonCheckbox, ButtonAssist };
+export { ButtonMain, ButtonCheckbox, ButtonAssist };
